fix: fix car-domain, car-util, race, console
